added hibernate dependency to get rid of annoying warning in startup
